# Conflicts:
#	Sources/Internal/Scene3D/Components/Waypoint/EdgeComponent.cpp
#	Sources/Internal/Scene3D/Entity.cpp
#	Sources/Internal/Scene3D/Entity.h
#	Sources/Internal/Scene3D/Scene.cpp
#	Tools/ResourceEditor/Classes/Deprecated/SceneValidator.cpp
#	Tools/ResourceEditor/Classes/Deprecated/SceneValidator.h
#	Tools/ResourceEditor/Classes/Qt/DockSceneTree/SceneTreeModel.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/BaseTransformProxies.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/CollisionSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/CollisionSystem/CollisionRenderObject.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/DebugDrawSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/EditorParticlesSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/ModifSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/PathSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/StructureSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/VisibilityCheckSystem/VisibilityCheckSystem.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/System/WayEditSystem.cpp


Former-commit-id: ff4d323af796c44d2873c82de47e074f31dc6206

+++ Sources/Internal/Base/AlignedAllocator.h
@@ -13,7 +13,7 @@ void FreeAlignedMemory(void*);
 template <class C, uint32 align, class... Args>
 C* CreateObjectAligned(Args&&... a)
 {
-    auto ptr = AllocateAlignedMemory(sizeof(C), align);
+    auto ptr = AllocateAlignedMemory(static_cast<uint32>(sizeof(C)), align);
     return new (ptr) C(std::forward<Args>(a)...);
 }
